update example with rnn
